% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{.validate_assay}
\alias{.validate_assay}
\title{Validate an assay name}
\usage{
.validate_assay(object, assay, allow_null = TRUE)
}
\arguments{
\item{object}{Seurat object.}

\item{assay}{Assay name.}

\item{allow_null}{Whether NULL is allowed.}
}
\value{
Invisibly returns TRUE.
}
\description{
Validate an assay name
}
\keyword{internal}
