% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{.validate_character_vector}
\alias{.validate_character_vector}
\title{Validate a character vector}
\usage{
.validate_character_vector(x, arg, allow_empty = FALSE)
}
\arguments{
\item{x}{Object to validate.}

\item{arg}{Argument name used in error messages.}

\item{allow_empty}{Whether an empty vector is allowed.}
}
\value{
Invisibly returns TRUE.
}
\description{
Validate a character vector
}
\keyword{internal}
