% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{.validate_features}
\alias{.validate_features}
\title{Validate feature names}
\usage{
.validate_features(object, features, assay = NULL, require_all = TRUE)
}
\arguments{
\item{object}{Seurat object.}

\item{features}{Character vector of feature names.}

\item{assay}{Assay name.}

\item{require_all}{Whether all features must exist.}
}
\value{
Invisibly returns TRUE.
}
\description{
Validate feature names
}
\keyword{internal}
