% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{.validate_logical}
\alias{.validate_logical}
\title{Validate a logical scalar}
\usage{
.validate_logical(x, arg, allow_null = FALSE)
}
\arguments{
\item{x}{Object to validate.}

\item{arg}{Argument name used in error messages.}

\item{allow_null}{Whether NULL is allowed.}
}
\value{
Invisibly returns TRUE.
}
\description{
Validate a logical scalar
}
\keyword{internal}
