% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{.validate_metadata_column}
\alias{.validate_metadata_column}
\title{Validate a metadata column}
\usage{
.validate_metadata_column(object, column, arg = "group.by", allow_null = TRUE)
}
\arguments{
\item{object}{Seurat object.}

\item{column}{Metadata column name.}

\item{arg}{Argument name used in error messages.}

\item{allow_null}{Whether NULL is allowed.}
}
\value{
Invisibly returns TRUE.
}
\description{
Validate a metadata column
}
\keyword{internal}
