% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{.validate_seurat_object}
\alias{.validate_seurat_object}
\title{Validate a Seurat object}
\usage{
.validate_seurat_object(object)
}
\arguments{
\item{object}{Object to validate.}
}
\value{
Invisibly returns TRUE.
}
\description{
Internal helper used by plotting functions.
}
\keyword{internal}
