% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_summary.R
\name{print.uenoy_review_summary}
\alias{print.uenoy_review_summary}
\title{Print a review summary}
\usage{
\method{print}{uenoy_review_summary}(x, ...)
}
\arguments{
\item{x}{A \code{uenoy_review_summary} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a review summary
}
