% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_registry.R
\name{read_annotation_registry}
\alias{read_annotation_registry}
\title{Read an annotation registry}
\usage{
read_annotation_registry(path, validate = TRUE)
}
\arguments{
\item{path}{Path to a CSV file.}

\item{validate}{Logical; validate after reading.}
}
\value{
A data frame.
}
\description{
Read an annotation registry
}
