% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ueno_blue_white_red.R
\name{ueno_blue_white_red}
\alias{ueno_blue_white_red}
\title{Blue-white-red color palette}
\usage{
ueno_blue_white_red(n)
}
\arguments{
\item{n}{A single integer greater than or equal to 2.}
}
\value{
A character vector of hexadecimal color codes.
}
\description{
Generate a blue-white-red color palette with the requested number
of colors.
}
\examples{
ueno_blue_white_red(11)
}
