% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{ueno_blue_white_red}
\alias{ueno_blue_white_red}
\title{Blue-white-red color palette}
\usage{
ueno_blue_white_red(
  n = 101L,
  low = "#0033FF",
  mid = "#FFFFFF",
  high = "#FF1A1A"
)
}
\arguments{
\item{n}{Number of colors to generate.}

\item{low}{Color used for low values.}

\item{mid}{Color used for midpoint values.}

\item{high}{Color used for high values.}
}
\value{
A character vector containing hexadecimal colors.
}
\description{
Returns a blue-white-red color palette for continuous expression values.
}
\examples{
ueno_blue_white_red()
ueno_blue_white_red(11)
}
